+++ market-batch/src/main/java/batch/server/alarm/application/MailEventHandler.java
@@ -1,6 +1,6 @@
 package batch.server.alarm.application;
 
-import batch.server.alarm.domain.MailStorage;
+import batch.server.alarm.domain.mail.MailStorage;
 import batch.server.alarm.domain.event.RegisteredEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
