+++ market-batch/src/main/java/batch/server/alarm/infrastructure/MailStorageJpaRepository.java
@@ -1,7 +1,7 @@
 package batch.server.alarm.infrastructure;
 
-import batch.server.alarm.domain.MailStatus;
-import batch.server.alarm.domain.MailStorage;
+import batch.server.alarm.domain.mail.vo.MailStatus;
+import batch.server.alarm.domain.mail.MailStorage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
